+++ test/JsonApiDotNetCoreTests/IntegrationTests/ResourceConstructorInjection/PostOffice.cs
@@ -7,7 +7,7 @@ namespace JsonApiDotNetCoreTests.IntegrationTests.ResourceConstructorInjection;
 
 [UsedImplicitly(ImplicitUseTargetFlags.Members)]
 [Resource(ControllerNamespace = "JsonApiDotNetCoreTests.IntegrationTests.ResourceConstructorInjection")]
-public sealed class PostOffice(InjectionDbContext injectionDbContext) : Identifiable<int>
+public sealed class PostOffice(InjectionDbContext injectionDbContext) : Identifiable<long>
 {
     private readonly TimeProvider _timeProvider = injectionDbContext.TimeProvider;
